requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
CommonMark==0.5.4
diff-match-patch==20121119
Django==2.0.3
django-import-export==1.2.0
django-summernote==0.8.8.6
docopt==0.6.2
docutils==0.14
et-xmlfile==1.0.1
Flask==0.12.2
idna==2.6
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
odfpy==1.3.6
openpyxl==2.4.9
path-and-address==2.0.1
Pygments==2.2.0
pytz==2018.3
PyYAML==3.12
recommonmark==0.4.0
requests==2.18.4
sphinx-rtd-theme==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
Werkzeug==0.14.1
xlrd==1.1.0
xlwt==1.3.0
